requirements.txt
python-dotenv
flask
mysqlclient
